refactor: consolidate skills-lab block into skills-editor
Made-with: Cursor
