--- nx/blocks/skills-lab/skills-lab.css
@@ -1,1 +1,0 @@
-/* Layout lives on da-browse-view (browse.css); block is a thin host. */
